Хаменила тип данных int на double

+++ src/Task_3.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <cmath>
 int main() {
-int start, end, step;
+double start, end, step;
 std::cout << "Введите начальное значение x: ";
 std::cin >> start; // Задаем исходное значение переменной
 std::cout << "Введите конечное значение x: ";
